requirements.txt
aniso8601==9.0.1
APScheduler==3.10.4
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.4.2
distlib==0.3.8
dnspython==2.6.1
exceptiongroup==1.2.0
filelock==3.13.1
Flask==3.0.2
Flask-RESTful==0.3.10
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mongoengine==0.27.0
mongomock==4.1.2
packaging==23.2
platformdirs==4.2.0
pluggy==1.4.0
pymongo==4.6.2
pytest==8.0.1
pytest-dotenv==0.5.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
sentinels==1.0.0
six==1.16.0
tomli==2.0.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
virtualenv==20.25.1
Werkzeug==3.0.1
zipp==3.17.0
